initial commit of repository refactoring
Git is hard. And learning it is not fun. So creating a new Repo instead of transforming the old one and preserving the history was easier to get working.

Already curious to see, if Thomas Levesque violates copyright again, rips off my codebase and gives 0 credit. What a glorious attitude that would be for a MVP